requirements.txt
asgiref==3.4.1
Django==3.2.5
django-cors-headers==3.7.0
django-mysql==3.12.0
djangorestframework==3.12.4
Faker==8.10.1
mysqlclient==2.0.3
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
